Get place closest to current location in search page. Remove unused location logic. Refactor rest of location logic.
