+++ src/services/geonames.service.ts
@@ -1,6 +1,6 @@
 import { GEONAMESUNIQUE } from '../../assets/geonames-unique';
 
-function placeFromGeoname(value: string): Geoname | undefined {
+function placeFromGeoname(value: string): Place | undefined {
   const fields = value.split(/\t/);
   const lat = parseFloat(fields[4]);
   const long = parseFloat(fields[5]);
@@ -9,8 +9,10 @@ function placeFromGeoname(value: string): Geoname | undefined {
   }
 
   return {
-    lat: parseFloat(fields[4]),
-    long: parseFloat(fields[5]),
+    position: {
+      lat: parseFloat(fields[4]),
+      long: parseFloat(fields[5]),
+    },
     name: fields[1],
   };
 }
@@ -19,11 +21,11 @@ function parseGeonames(geonames: string): Array<string> {
   return geonames.split(/\r?\n/);
 }
 
-export function getGeonames(): Record<string, Geoname> {
+export function getGeonames(): Record<string, Place> {
   const rawPlaces = parseGeonames(GEONAMESUNIQUE);
 
   return rawPlaces.reduce(
-    (acc: Record<string, Geoname>, val) => {
+    (acc: Record<string, Place>, val) => {
       const place = placeFromGeoname(val);
       if (place) {
         ((acc[place.name] = place), acc)
