+++ src/store/location.slice.ts
@@ -1,71 +1,11 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { requestMultiple, PERMISSIONS, Permission, openSettings, checkMultiple } from 'react-native-permissions';
-import Geolocation from '@react-native-community/geolocation';
-import { Platform } from 'react-native';
 
 import { DISTRICTS } from '../constants/districts.constant';
 import { LOGGER } from '../lib';
-import { snapToPlace } from '../utils/places';
-
-const checkPermissions = async (permissions: Array<Permission>): Promise<boolean> => {
-  const statuses = await checkMultiple(permissions);
-  return Object.values(statuses).reduce((acc, val) => acc && val === 'granted', true);
-}
-
-const requestPermissions = async (permissions: Array<Permission>): Promise<boolean> => {
-  const alreadyGranted = await checkPermissions(permissions);
-  if (alreadyGranted) {
-    return true;
-  }
-
-  const requests = await requestMultiple(permissions);
-  const granted = Object.values(requests).reduce((acc, val) => acc && val === 'granted', true);
-  if (!granted) {
-    const errors = Object.entries(requests).map(([k, v]) => `${k}: ${v}`).join('\n');
-    throw new Error(errors);
-  }
-
-  return true;
-}
-
-const requestAndroid = async (): Promise<boolean> => {
-  const needed = [PERMISSIONS.ANDROID.ACCESS_COARSE_LOCATION, PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION];
-  return await requestPermissions(needed);
-};
-
-const requestIOS = async () => {
-  const needed = [PERMISSIONS.IOS.LOCATION_ALWAYS, PERMISSIONS.IOS.LOCATION_WHEN_IN_USE];
-  return await requestPermissions(needed);
-};
-
-function queryLocation(): Promise<{ lat: number, lon: number }> {
-  return new Promise((resolve, reject) => {
-    Geolocation.getCurrentPosition(
-      position =>
-        resolve({
-          lat: position.coords.latitude,
-          lon: position.coords.longitude,
-        }),
-      error => reject(error),
-      { enableHighAccuracy: true },
-    );
-  });
-}
-
-const getClosestPlace = (lat: number, lon: number): Place|undefined => {
-  const place: Place = { name: '', position: { lat, long: lon } };
-  return snapToPlace(place);
-}
+import { placeByCurrentLocation } from '../utils/location';
 
 export const getPreciseLocation = createAsyncThunk('location/getPreciseLocation', async (): Promise<Place> => {
-  Platform.OS === 'ios' ? await requestIOS() : await requestAndroid();
-  const location = await queryLocation();
-  const closest = getClosestPlace(location.lat, location.lon);
-  if (!closest) {
-    throw new Error('Failed to get closest location.');
-  }
-
-  return closest as Place;
+  return placeByCurrentLocation()
 });
 
 type InitialState = {
@@ -96,11 +36,11 @@ const locationSlice = createSlice({
       state.loading = true;
     });
     builder.addCase(getPreciseLocation.fulfilled, (state, action) => {
-      LOGGER.info('Loading location fulfilled.');
-      state.loading = false;
-      state.name = action.payload.name;
-      state.lat = action.payload.position.lat;
-      state.lon = action.payload.position.long;
+        LOGGER.info('Loading location fulfilled.');
+        state.loading = false;
+        state.name = action.payload.name;
+        state.lat = action.payload.position.lat;
+        state.lon = action.payload.position.long;
     });
     builder.addCase(getPreciseLocation.rejected, (state, action) => {
       LOGGER.error('Loading location rejected.');
